+++ internal/config/config/config.go
@@ -18,6 +18,7 @@ var keywords = set.NewSet[string](
 	"version",
 	"shell",
 	"env",
+	"env_file",
 	"init",
 	"before",
 	"mixins",
@@ -35,16 +36,20 @@ type Config struct {
 	// before is a script which will be included before every cmd
 	Before string
 	// init is a script which will be called exactly once before any command calls
-	Init    string
-	Env     *Envs
-	Version string
-	isMixin bool // if true, we consider config as mixin and apply different parsing and validation
+	Init     string
+	Env      *Envs
+	EnvFiles *EnvFiles
+	Version  string
 	// absolute path to .lets
 	DotLetsDir string
 	// absolute path to .lets/checksums
 	ChecksumsDir string
 	// absolute path to .lets/mixins
 	MixinsDir string
+
+	// cached env after config.SetupEnv, used in config.GetEnv
+	cachedEnv map[string]string
+	isMixin   bool // if true, we consider config as mixin and apply different parsing and validation
 }
 
 func (c *Config) UnmarshalYAML(unmarshal func(interface{}) error) error {
@@ -68,6 +73,7 @@ func (c *Config) UnmarshalYAML(unmarshal func(interface{}) error) error {
 		Before   string
 		Init     string
 		Env      *Envs
+		EnvFiles *EnvFiles `yaml:"env_file"`
 	}
 
 	if err := unmarshal(&config); err != nil {
@@ -91,6 +97,10 @@ func (c *Config) UnmarshalYAML(unmarshal func(interface{}) error) error {
 	c.Env = config.Env
 	if c.Env == nil {
 		c.Env = &Envs{}
+	}
+	c.EnvFiles = config.EnvFiles
+	if c.EnvFiles == nil {
+		c.EnvFiles = &EnvFiles{}
 	}
 
 	for name, cmd := range c.Commands {
@@ -188,6 +198,7 @@ func (c *Config) mergeMixin(mixin *Config) error {
 		c.Before,
 		mixin.Before,
 	)
+	c.EnvFiles.Append(mixin.EnvFiles)
 
 	return nil
 }
@@ -281,6 +292,11 @@ func (c *Config) readMixins(mixins []*Mixin) error {
 }
 
 func (c *Config) GetEnv() map[string]string {
+	if c.cachedEnv != nil {
+		// clone to avoid mutating config env
+		return cloneMap(c.cachedEnv)
+	}
+
 	return c.Env.Dump()
 }
 
@@ -291,12 +307,28 @@ func (c *Config) SetupEnv() error {
 		return err
 	}
 
+	filenameEnv := c.BuiltinEnv(c.Shell)
+	for key, value := range c.Env.Dump() {
+		filenameEnv[key] = value
+	}
+
+	envFileEnv, err := c.EnvFiles.Load(*c, filenameEnv)
+	if err != nil {
+		return fmt.Errorf("failed to resolve global env_file: %w", err)
+	}
+
+	c.cachedEnv = c.Env.Dump()
+	for key, value := range envFileEnv {
+		c.cachedEnv[key] = value
+	}
+
 	// expand env for args
 	for _, cmd := range c.Commands {
 		for idx, arg := range cmd.Args {
 			// we have to expand env here on our own, since this args not came from users tty, and not expanded before lets
 			cmd.Args[idx] = os.Expand(arg, func(key string) string {
-				return c.Env.Mapping[key].Value
+				// safe to access own cached env
+				return c.cachedEnv[key]
 			})
 		}
 	}